requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
beautifultable==1.0.1
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
ecdsa==0.17.0
fastapi==0.66.0
filelock==3.0.12
fire==0.4.0
h11==0.12.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pysondb==1.1.4
python-jose==3.3.0
rsa==4.7.2
six==1.16.0
starlette==0.14.2
termcolor==1.1.0
typing-extensions==3.10.0.0
uvicorn==0.14.0
wcwidth==0.2.5
